systemd-conf: Rename dwmac to ethsom0 and ksz8851 to ethsom1
The DHCOR SoM has one ethernet interface, the DWMAC4, by default eth0.
The DHCOM SoM has two ethernet interfaces, the DWMAC4 and KS8851, by
default eth1 and eth0 respectively. The reversal of eth1 and eth0 is
due to the undefined order in which Linux kernel probes devices, hence
the kernel interface names are assigned as the interfaces come up. On
the DHCOM, the KS8851 is probed before the DWMAC. By adding additional
ethernet interfaces into the system, the ethX assignment might change
even further, which means the userspace should only depend on stable
interface identifiers, like platform path, PCI topology path, etc.

To make the naming of the integrated interfaces consistent and stable,
this patch adds systemd link files for all DHSOM modules and renames
the integrated interfaces as follows:
  DWMAC4 always ethsom0
  KS8851 always ethsom1
The names are matched on platform path and hence do not change unless
the interface is physically relocated. For DWMAC4 that means a change
in the SoC itself, which is highly unlikely. For KS8851 that might be
a move of the KS8851 to different chipselect within the EBI controller.

Note that it is not possible to simply swap eth0 and eth1 names due to
naming collision, hence unique ethernet device names are used instead.

Signed-off-by: Marek Vasut <marex@denx.de>