systemd-conf: Prefer descrete RTC on the SoM over RTC built into the SoC
The discrete RTC chip on the SoM is higher precision and often battery-backed,
prefer it over the RTC built into the SoC itself, which is not battery-backed.

By default, udev creates a symlink from the first RTC device node to /dev/rtc,
which is then used by e.g. systemd-timesyncd to set system date and time. The
RTC device node enumeration happens in the order in which the RTC drivers get
probed, and that order may change, e.g. if an RTC driver is turned into module
or if a new RTC is added into the system.

To define preference for the discrete RTC chip, udev symlink priorities are
used. The default udev rule which describes /dev/rtc symlink matches on all
RTCs and sets the symlink priority to -100. Any subsequent udev rule which
matches and sets higher symlink priority will override the symlink. Therefore
this patch adds a rule which matches on RTC with "name" attribute "rtc-rv?*",
which is all RTCs from MC RV* line up (e.g. RV8803) and sets the symlink prio
to -99, which is higher than default -100.

Signed-off-by: Marek Vasut <marex@denx.de>