image: Replace canutils with can-utils
The meta-openembedded layer currently has two variants of can utilities,
where the recipe name differs only in a single dash "-" character. The
variant without a dash is a legacy downstream fork retained likely only
for the sake of backward compatibility, the variant with a dash is the
upstream version, used also e.g. in debian. Replace the downstream fork
with upstream version.

Signed-off-by: Marek Vasut <marex@denx.de>