weston-init: Import weston.ini from oe-core
Import weston.ini copy from oe-core branch dunfell as of commit
cffe25dcbd ("weston-init: Fix weston-keyboard path in weston.ini")

There are no changes to this file thus far. This file is currently
picked as a replacement of the file present in oe-core in case the
dh-dhsom is present in MACHINEOVERRIDES , which is on all DHSOM
derivatives.

Signed-off-by: Marek Vasut <marex@denx.de>