opengl-es-cts: Add kirkstone specific bbappend in par with dunfell one
The up-to-date opengl-es-cts in OE dunfell release is present in
meta-mainline-graphics, which is not compatible with OE kirkstone,
while no opengl-es-cts exists in upstream meta-openembedded layer
dunfell branch yet. The up-to-date opengl-es-cts in OE kirkstone
release is already present in meta-openembedded kirkstone branch.

Handle this by introducing two opengl-es-cts_%.bbappend, one for
each OE release, in their respective metalayer dynamic layers.
Include those dynamic layers in the build only in case of matching
OE release. This is necessary to avoid build error if meta-openembedded
is included in the OE dunfell release build, which would attempt to
apply opengl-es-cts_%.bbappend on opengl-es-cts base recipe missing
from meta-openembedded dunfell branch.

Note that this already supports opengl-es-cts present in OE kirkstone
release and newer, the match is implemented as a test for OE dunfell
release and apply compatibility bbappend instead.

Signed-off-by: Marek Vasut <marex@denx.de>