layer: Add downstream BSP support for MX8MP DHCOM PDK2
Add support for meta-imx and meta-freescale which provide downstream BSP
for the MX8MP DHCOM PDK2. This downstream BSP can be used to operate the
NPU until there is a suitable upstream driver. Keep in mind the downstream
BSP is seldom receiving any security updates and contains vast amount of
unreviewed code.

Signed-off-by: Marek Vasut <marex@denx.de>