dynamic/flutter: flutter-engine: Create temporary missing files
The meta-flutter builds currently fail on do_install of flutter-engine
on 32bit systems due to missing analyze_snapshot. This is expected, as
the binary is only generated for 64bit builds, while meta-flutter does
attempt to install it unconditionally. As the flutter-engine do_install
cannot be easily updated, create substitute analyze_snapshot in ${S} .
Use the same approach for dart, where it is unclear why this should
ever be installed on the target.

Signed-off-by: Marek Vasut <marex@denx.de>