gdk-pixbuf: Re-enable fringe BMP loader
The gdk-pixbuf disabled fringe loaders, like BMP, by default in 2.42.11
release. The rationale is that this makes it easier for distros to ship
a sane subset of loaders. The dh-image-demo ships with BMP loader support
for gstreamer gdkpixbuf use, reenable the 'others' loaders as the BMP
loader cannot be enabled separately.

Signed-off-by: Marek Vasut <marex@denx.de>