util-linux: Fix pointer to pointer misuse in ul_strtou64()
util-linux 2.39.4 and older in OE scarthgap passes pointer to pointer into
ul_strtou64(), instead of pointer to memory. Adjust the code to pass pointer
to memory. The fix might be part of util-linux 2.39.5 if it is ever released,
at which point the sed invocation will have no effect on the source code and
could be dropped. Upstream util-linux fix commit is 1064a53e4ff3 ("sys-utils:
hwclock-rtc: fix pointer usage") .

Signed-off-by: Marek Vasut <marex@nabladev.com>