systemd-conf: Update PHY LED udev rules for STM32MP13xx DHCOR DHSBC
Linux commit 52c47742f79d ("leds: triggers: send uevent when changing triggers")
clarifies that when a LED trigger is configured via sysfs attribute, the new
attributes instantiated by that trigger are not immediately created. When the
attributes are created, a "change" event is emitted, which allows userspace to
configure file permissions of those new attributes. Use the same change event
to configure the content of those new trigger attributes for PHY LEDs.

This mitigates the problem, where some of the attributes were not assigned
during the "add" event, because they were not yet instantiated.

Signed-off-by: Marek Vasut <marex@nabladev.com>